requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
clusteval==2.2.1
colourmap==1.1.17
contourpy==1.2.1
cycler==0.12.1
datazets==0.1.9
fonttools==4.53.0
idna==3.7
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.0
mysql-connector-python==8.4.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
pyparsing==3.1.2
pypickle==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scatterd==1.3.7
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
tenacity==8.4.1
threadpoolctl==3.5.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.2
